style: CreatedEventList 컴포넌트 분리
