+++ client/src/pages/mypage/events/CreatedEventsPage.tsx
@@ -2,20 +2,11 @@ import {css} from '@emotion/react';
 import {useEffect, useState} from 'react';
 
 import useRequestGetCreatedEvents from '@hooks/queries/event/useRequestGetCreatedEvents';
-import {CreatedEvent} from 'types/serviceType';
-import {CreatedEventItem} from '@components/Design/components/CreatedEvent/CreatedEvent';
-import useRequestDeleteEvents from '@hooks/queries/event/useRequestDeleteEvents';
 
-import {FixedButton, Flex, FunnelLayout, Input, MainLayout, TextButton, Top, TopNav} from '@components/Design';
+import {FunnelLayout, MainLayout, TextButton, Top, TopNav} from '@components/Design';
+import {CreatedEventList} from '@components/CreatedEventList';
 
 import {useCreatedEventsPageContext, CreatedEventsPageContextProvider} from './CreatedEvent.context';
-
-interface CreatedEventListProps {
-  eventName: string;
-  onSearch: ({target}: React.ChangeEvent<HTMLInputElement>) => void;
-  placeholder: string;
-  createdEvents: CreatedEvent[];
-}
 
 const PageInner = () => {
   const [eventName, setEventName] = useState('');
@@ -71,51 +62,5 @@ export default function CreatedEventsPage() {
         <PageInner />
       </CreatedEventsPageContextProvider>
     </MainLayout>
-  );
-}
-
-function CreatedEventList({createdEvents, eventName, onSearch, placeholder}: CreatedEventListProps) {
-  const {mode, handleMode, selectedEvents, has, handleSelectedEvents} = useCreatedEventsPageContext();
-  const setViewMode = () => handleMode('view');
-
-  const {deleteEvents} = useRequestDeleteEvents();
-
-  const onDeleteClick = async () => {
-    const selectedEventsId = selectedEvents.map(event => event.eventId);
-    await deleteEvents({eventIds: selectedEventsId});
-    handleMode('view');
-  };
-
-  return (
-    <>
-      <Flex
-        flexDirection="column"
-        width="100%"
-        backgroundColor="white"
-        padding="0.5rem 1rem"
-        paddingInline="0.5rem"
-        gap="0.5rem"
-        height="100%"
-        cssProp={{borderRadius: '1rem'}}
-      >
-        <Input inputType="search" value={eventName} onChange={onSearch} placeholder={placeholder} />
-        {createdEvents.length !== 0 &&
-          createdEvents.map(createdEvent => (
-            <CreatedEventItem
-              key={createdEvent.eventId}
-              isEditMode={mode === 'edit'}
-              setEditMode={() => handleMode('edit')}
-              isChecked={has(createdEvent)}
-              onChange={handleSelectedEvents}
-              createdEvent={createdEvent}
-            />
-          ))}
-      </Flex>
-      {mode === 'edit' && (
-        <FixedButton variants="tertiary" onDeleteClick={onDeleteClick} onClick={setViewMode}>
-          편집완료
-        </FixedButton>
-      )}
-    </>
   );
 }
